+++ core/deferredGraphics/renderStages/source/lighting.cpp
@@ -89,7 +89,7 @@ void graphics::createLightingDescriptorSets()
 
 void graphics::updateLightingDescriptorSets(camera* cameraObject)
 {
-    for (size_t i = 0; i < image.Count; i++)
+    for (uint32_t i = 0; i < image.Count; i++)
     {
         std::vector<VkDescriptorImageInfo> imageInfos;
         for(uint32_t index = 0; index < 4;index++)
